+++ comet_pqc/preferences.py
@@ -132,6 +132,19 @@ class TableTab(PreferencesTab):
             suffix="mm",
             changed=self.on_z_limit_movement_changed
         )
+        def make_number():
+            return ui.Number(
+                minimum=0,
+                maximum=1000.0,
+                decimals=3,
+                suffix="mm"
+            )
+        self.probecard_limit_x_maximum_number = make_number()
+        self.probecard_limit_y_maximum_number = make_number()
+        self.probecard_limit_z_maximum_number = make_number()
+        self.joystick_limit_x_maximum_number = make_number()
+        self.joystick_limit_y_maximum_number = make_number()
+        self.joystick_limit_z_maximum_number = make_number()
         self.layout = ui.Column(
             ui.GroupBox(
                 title="Control Steps (mm)",
@@ -147,11 +160,56 @@ class TableTab(PreferencesTab):
                 )
             ),
             ui.GroupBox(
-                title="Movement Z-Limit",
+            title="Movement Z-Limit",
                 layout=ui.Column(
                     self.z_limit_movement_number
                 )
-            )
+            ),
+            ui.GroupBox(
+                title="Probe Card Limts",
+                layout=ui.Row(
+                    ui.Column(
+                        ui.Label("X"),
+                        self.probecard_limit_x_maximum_number
+                    ),
+                    ui.Column(
+                        ui.Label("Y"),
+                        self.probecard_limit_y_maximum_number
+                    ),
+                    ui.Column(
+                        ui.Label("Z"),
+                        self.probecard_limit_z_maximum_number
+                    ),
+                    ui.Column(
+                        ui.Label(),
+                        ui.Label("Maximum"),
+                    ),
+                    ui.Spacer()
+                )
+            ),
+            ui.GroupBox(
+                title="Joystick Limits",
+                layout=ui.Row(
+                    ui.Column(
+                        ui.Label("X"),
+                        self.joystick_limit_x_maximum_number
+                    ),
+                    ui.Column(
+                        ui.Label("Y"),
+                        self.joystick_limit_y_maximum_number
+                    ),
+                    ui.Column(
+                        ui.Label("Z"),
+                        self.joystick_limit_z_maximum_number
+                    ),
+                    ui.Column(
+                        ui.Label(),
+                        ui.Label("Maximum"),
+                    ),
+                    ui.Spacer()
+                )
+            ),
+            stretch=(1, 0, 0, 0)
         )
 
     def on_position_selected(self, item):
@@ -208,6 +266,16 @@ class TableTab(PreferencesTab):
             ))
         self.steps_tree.qt.sortByColumn(0, QtCore.Qt.AscendingOrder)
         self.z_limit_movement_number.value = settings.table_z_limit
+        # Probecard limits
+        x, y, z = settings.table_probecard_maximum_limits
+        self.probecard_limit_x_maximum_number.value = x
+        self.probecard_limit_y_maximum_number.value = y
+        self.probecard_limit_z_maximum_number.value = z
+        # Joystick limits
+        x, y, z = settings.table_joystick_maximum_limits
+        self.joystick_limit_x_maximum_number.value = x
+        self.joystick_limit_y_maximum_number.value = y
+        self.joystick_limit_z_maximum_number.value = z
 
     def store(self):
         table_step_sizes = []
@@ -219,6 +287,18 @@ class TableTab(PreferencesTab):
             ))
         self.settings['table_step_sizes'] = table_step_sizes
         settings.table_z_limit = self.z_limit_movement_number.value
+        # Probecard limits
+        settings.table_probecard_maximum_limits = [
+            self.probecard_limit_x_maximum_number.value,
+            self.probecard_limit_y_maximum_number.value,
+            self.probecard_limit_z_maximum_number.value
+        ]
+        # Joystick limits
+        settings.table_joystick_maximum_limits = [
+            self.joystick_limit_x_maximum_number.value,
+            self.joystick_limit_y_maximum_number.value,
+            self.joystick_limit_z_maximum_number.value
+        ]
 
 class OptionsTab(PreferencesTab):
     """Options tab for preferences dialog."""
@@ -226,6 +306,8 @@ class OptionsTab(PreferencesTab):
     def __init__(self):
         super().__init__(title="Options")
         self.png_plots_checkbox = ui.CheckBox("Save plots as PNG")
+        self.points_in_plots_checkbox = ui.CheckBox("Show points in plots")
+        self.png_analysis_checkbox = ui.CheckBox("Add analysis preview to PNG")
         self.export_json_checkbox = ui.CheckBox("Write JSON data (*.json)")
         self.export_txt_checkbox = ui.CheckBox("Write plain text data (*.txt)")
         self.write_logfiles_checkbox = ui.CheckBox("Write measurement log files (*.log)")
@@ -234,8 +316,15 @@ class OptionsTab(PreferencesTab):
         self.layout = ui.Column(
             ui.GroupBox(
                 title="Plots",
-                layout=ui.Row(
-                    self.png_plots_checkbox
+                layout=ui.Column(
+                    self.png_plots_checkbox,
+                    self.points_in_plots_checkbox,
+                )
+            ),
+            ui.GroupBox(
+                title="Analysis",
+                layout=ui.Column(
+                    self.png_analysis_checkbox
                 )
             ),
             ui.GroupBox(
@@ -270,13 +359,17 @@ class OptionsTab(PreferencesTab):
         )
 
     def load(self):
-        png_plots = self.settings.get("png_plots", False)
+        png_plots = self.settings.get("png_plots") or False
         self.png_plots_checkbox.checked = png_plots
-        export_json = self.settings.get("export_json", False)
+        points_in_plots = self.settings.get("points_in_plots") or False
+        self.points_in_plots_checkbox.checked = points_in_plots
+        png_analysis = self.settings.get("png_analysis") or False
+        self.png_analysis_checkbox.checked = png_analysis
+        export_json = self.settings.get("export_json") or False
         self.export_json_checkbox.checked = export_json
-        export_txt = self.settings.get("export_txt", True)
+        export_txt = self.settings.get("export_txt") or True
         self.export_txt_checkbox.checked = export_txt
-        write_logfiles = self.settings.get("write_logfiles", True)
+        write_logfiles = self.settings.get("write_logfiles") or True
         self.write_logfiles_checkbox.checked = write_logfiles
         vsrc_instrument = self.settings.get("vsrc_instrument") or "K2657A"
         if vsrc_instrument in self._vsrc_instrument_combobox:
@@ -288,6 +381,10 @@ class OptionsTab(PreferencesTab):
     def store(self):
         png_plots = self.png_plots_checkbox.checked
         self.settings["png_plots"] = png_plots
+        points_in_plots = self.points_in_plots_checkbox.checked
+        self.settings["points_in_plots"] = points_in_plots
+        png_analysis = self.png_analysis_checkbox.checked
+        self.settings["png_analysis"] = png_analysis
         export_json = self.export_json_checkbox.checked
         self.settings["export_json"] = export_json
         export_txt = self.export_txt_checkbox.checked