+++ comet_pqc/tablecontrol.py
@@ -11,9 +11,12 @@ from comet_pqc.utils import from_table_unit, to_table_unit
 from comet_pqc.utils import format_metric
 
 from .components import PositionLabel
+from .components import PositionWidget
+from .components import CalibrationWidget
 from .components import ToggleButton
 from .settings import settings, TablePosition
-from .utils import format_switch
+from .utils import format_switch, caldone_valid
+from .position import Position
 
 from qutie.qutie import Qt
 
@@ -113,8 +116,8 @@ class TableContactItem(ui.TreeItem):
 
 class TableContactsWidget(ui.Row):
 
-    def __init__(self, position_picked=None, absolute_move=None):
-        super().__init__()
+    def __init__(self, position_picked=None, absolute_move=None, **kwargs):
+        super().__init__(**kwargs)
         self.position_picked = position_picked
         self.absolute_move = absolute_move
         self.contacts_tree = ui.Tree(
@@ -201,7 +204,7 @@ class TableContactsWidget(ui.Row):
             if current_item.has_position:
                 if ui.show_question(f"Do you want to move tabel to contact {current_item.name}?"):
                     x, y, z = current_item.position
-                    self.emit(self.absolute_move, x, y, z)
+                    self.emit(self.absolute_move, Position(x, y, z))
 
     def on_calculate(self):
         current_item = self.contacts_tree.current
@@ -269,62 +272,88 @@ class TablePositionItem(ui.TreeItem):
 
 class PositionDialog(ui.Dialog):
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.name_text = ui.Text(value="Unnamed")
-        self.x_number = ui.Number(value=0., minimum=0., maximum=1000., decimals=3, suffix="mm")
-        self.y_number = ui.Number(value=0., minimum=0., maximum=1000., decimals=3, suffix="mm")
-        self.z_number = ui.Number(value=0., minimum=0., maximum=1000., decimals=3, suffix="mm")
-        self.comment_text = ui.Text()
-        self.button_box = ui.DialogButtonBox(buttons=("ok", "cancel"), accepted=self.accept, rejected=self.reject)
+    def __init__(self, position_picked=None, **kwargs):
+        super().__init__(**kwargs)
+        self.position_picked = position_picked
+        self._name_text = ui.Text(value="Unnamed")
+        self._x_number = ui.Number(value=0., minimum=0., maximum=1000., decimals=3, suffix="mm")
+        self._y_number = ui.Number(value=0., minimum=0., maximum=1000., decimals=3, suffix="mm")
+        self._z_number = ui.Number(value=0., minimum=0., maximum=1000., decimals=3, suffix="mm")
+        self._comment_text = ui.Text()
+        self._assign_button = ui.Button(
+            text="Assign Position",
+            tool_tip="Assign current table position.",
+            clicked=self.on_assign_clicked
+        )
+        self._button_box = ui.DialogButtonBox(buttons=("ok", "cancel"), accepted=self.accept, rejected=self.reject)
         self.layout = ui.Column(
             ui.Label("Name", tool_tip="Position name"),
-            self.name_text,
-            ui.Label("X", tool_tip="Position X coordinate"),
-            self.x_number,
-            ui.Label("Y", tool_tip="Position Y coordinate"),
-            self.y_number,
-            ui.Label("Z", tool_tip="Position Z coordinate"),
-            self.z_number,
+            self._name_text,
+            ui.Row(
+                ui.Column(
+                    ui.Label("X", tool_tip="Position X coordinate"),
+                    self._x_number
+                ),
+                ui.Column(
+                    ui.Label("Y", tool_tip="Position Y coordinate"),
+                    self._y_number
+                ),
+                ui.Column(
+                    ui.Label("Z", tool_tip="Position Z coordinate"),
+                    self._z_number
+                ),
+                ui.Column(
+                    ui.Spacer(),
+                    self._assign_button,
+                )
+            ),
             ui.Label("Comment", tool_tip="Optional position comment"),
-            self.comment_text,
-            self.button_box
+            self._comment_text,
+            ui.Spacer(),
+            self._button_box
         )
 
     @property
     def name(self):
-        return self.name_text.value
+        return self._name_text.value
 
     @name.setter
     def name(self, value):
-        self.name_text.value = value
+        self._name_text.value = value
 
     @property
     def position(self):
-        x = self.x_number.value
-        y = self.y_number.value
-        z = self.z_number.value
+        x = self._x_number.value
+        y = self._y_number.value
+        z = self._z_number.value
         return x, y, z
 
     @position.setter
     def position(self, value):
         x, y, z = value
-        self.x_number.value = x
-        self.y_number.value = y
-        self.z_number.value = z
+        self._x_number.value = x
+        self._y_number.value = y
+        self._z_number.value = z
 
     @property
     def comment(self):
-        return self.comment_text.value
+        return self._comment_text.value
 
     @comment.setter
     def comment(self, value):
-        self.comment_text.value = value
+        self._comment_text.value = value
+
+    def on_assign_clicked(self):
+        self.layout.enabled = False
+        def callback(x, y, z):
+            self.position = x, y, z
+            self.layout.enabled = True
+        self.emit(self.position_picked, callback)
 
 class TablePositionsWidget(ui.Row, SettingsMixin):
 
-    def __init__(self, position_picked=None, absolute_move=None):
-        super().__init__()
+    def __init__(self, position_picked=None, absolute_move=None, **kwargs):
+        super().__init__(**kwargs)
         self.position_picked = position_picked
         self.absolute_move = absolute_move
         self.positions_tree = ui.Tree(
@@ -332,12 +361,6 @@ class TablePositionsWidget(ui.Row, SettingsMixin):
             root_is_decorated=False,
             selected=self.on_position_selected,
             double_clicked=self.on_position_double_clicked
-        )
-        self.pick_button = ui.Button(
-            text="Assign &Position",
-            tool_tip="Assign current table position to selected position item",
-            clicked=self.on_pick_position,
-            enabled=False
         )
         self.add_button = ui.Button(
             text="&Add",
@@ -364,7 +387,6 @@ class TablePositionsWidget(ui.Row, SettingsMixin):
         )
         self.append(self.positions_tree)
         self.append(ui.Column(
-            self.pick_button,
             self.move_button,
             ui.Spacer(),
             self.add_button,
@@ -399,7 +421,6 @@ class TablePositionsWidget(ui.Row, SettingsMixin):
         settings.table_positions = positions
 
     def lock(self):
-        self.pick_button.enabled = False
         self.add_button.enabled = False
         self.edit_button.enabled = False
         self.remove_button.enabled = False
@@ -409,7 +430,6 @@ class TablePositionsWidget(ui.Row, SettingsMixin):
 
     def unlock(self):
         enabled = self.positions_tree.current is not None
-        self.pick_button.enabled = enabled
         self.add_button.enabled = True
         self.edit_button.enabled = enabled
         self.remove_button.enabled = enabled
@@ -419,7 +439,6 @@ class TablePositionsWidget(ui.Row, SettingsMixin):
 
     def on_position_selected(self, item):
         enabled = item is not None
-        self.pick_button.enabled = True
         self.edit_button.enabled = True
         self.remove_button.enabled = True
         self.move_button.enabled = True
@@ -427,17 +446,13 @@ class TablePositionsWidget(ui.Row, SettingsMixin):
     def on_position_double_clicked(self, *args):
         self.on_move()
 
-    def on_pick_position(self):
-        item = self.positions_tree.current
-        if item:
-            if ui.show_question(f"Do you want to assign current position to '{item.name}'?"):
-                def callback(x, y, z):
-                    item.position = x, y, z
-                    self.positions_tree.fit()
-                self.emit(self.position_picked, callback)
+    def on_position_picked(self, callback):
+        self.emit(self.position_picked, callback)
 
     def on_add_position(self):
-        dialog = PositionDialog()
+        dialog = PositionDialog(
+            position_picked=self.on_position_picked
+        )
         if dialog.run():
             name = dialog.name
             x, y, z = dialog.position
@@ -448,7 +463,9 @@ class TablePositionsWidget(ui.Row, SettingsMixin):
     def on_edit_position(self):
         item = self.positions_tree.current
         if item:
-            dialog = PositionDialog()
+            dialog = PositionDialog(
+                position_picked=self.on_position_picked
+            )
             dialog.name = item.name
             dialog.position = item.position
             dialog.comment = item.comment
@@ -464,7 +481,6 @@ class TablePositionsWidget(ui.Row, SettingsMixin):
             if ui.show_question(f"Do you want to remove position '{item.name}'?"):
                 self.positions_tree.remove(item)
                 if not len(self.positions_tree):
-                    self.pick_button.enabled = False
                     self.edit_button.enabled = False
                     self.remove_button.enabled = False
                 self.positions_tree.fit()
@@ -474,7 +490,7 @@ class TablePositionsWidget(ui.Row, SettingsMixin):
         if item:
             if ui.show_question(f"Do you want to move table to position '{item.name}'?"):
                 x, y ,z = item.position
-                self.emit(self.absolute_move, x, y, z)
+                self.emit(self.absolute_move, Position(x, y, z))
 
 class TableControlDialog(ui.Dialog, SettingsMixin):
 
@@ -602,23 +618,21 @@ class TableControlDialog(ui.Dialog, SettingsMixin):
             stretch=(0, 1)
         )
         self.positions_widget = TablePositionsWidget(
+            enabled=False,
             position_picked=self.on_position_picked,
             absolute_move=self.on_absolute_move
         )
         self.contacts_widget = TableContactsWidget(
+            enabled=False,
             position_picked=self.on_position_picked,
             absolute_move=self.on_absolute_move
         )
-        self.pos_x_label = PositionLabel()
-        self.pos_y_label = PositionLabel()
-        self.pos_z_label = PositionLabel()
-        self.cal_x_label = CalibrationLabel("cal")
-        self.cal_y_label = CalibrationLabel("cal")
-        self.cal_z_label = CalibrationLabel("cal")
-        self.rm_x_label = CalibrationLabel("rm")
-        self.rm_y_label = CalibrationLabel("rm")
-        self.rm_z_label = CalibrationLabel("rm")
+        self._position_widget = PositionWidget()
+        self._calibration_widget = CalibrationWidget()
         self.z_limit_label = PositionLabel()
+        self.x_hard_limit_label = PositionLabel()
+        self.y_hard_limit_label = PositionLabel()
+        self.z_hard_limit_label = PositionLabel()
         self.laser_label = SwitchLabel()
         self.calibrate_button = ui.Button(
             text="Calibrate",
@@ -723,50 +737,31 @@ class TableControlDialog(ui.Dialog, SettingsMixin):
                     stretch=(0, 1)
                 ),
                 ui.Column(
+                    self._position_widget,
+                    self._calibration_widget,
                     ui.GroupBox(
-                        title="Position",
-                        layout=ui.Row(
-                            ui.Column(
-                                ui.Label("X"),
-                                ui.Label("Y"),
-                                ui.Label("Z")
-                            ),
-                            ui.Column(
-                                self.pos_x_label,
-                                self.pos_y_label,
-                                self.pos_z_label
-                            )
-                        )
-                    ),
-                    ui.GroupBox(
-                        title="Calibration",
-                        layout=ui.Row(
-                            ui.Column(
-                                ui.Label("X"),
-                                ui.Label("Y"),
-                                ui.Label("Z")
-                            ),
-                            ui.Column(
-                                self.cal_x_label,
-                                self.cal_y_label,
-                                self.cal_z_label
-                            ),
-                            ui.Column(
-                                self.rm_x_label,
-                                self.rm_y_label,
-                                self.rm_z_label
-                            ),
-                            stretch=(1, 1)
-                        )
-                    ),
-                    ui.GroupBox(
-                        title="Limits",
+                        title="Soft Limits",
                         layout=ui.Row(
                             ui.Column(
                                 ui.Label("Z")
                             ),
                             ui.Column(
                                 self.z_limit_label
+                            )
+                        )
+                    ),
+                    ui.GroupBox(
+                        title="Hard Limits",
+                        layout=ui.Row(
+                            ui.Column(
+                                ui.Label("X"),
+                                ui.Label("Y"),
+                                ui.Label("Z")
+                            ),
+                            ui.Column(
+                                self.x_hard_limit_label,
+                                self.y_hard_limit_label,
+                                self.z_hard_limit_label
                             )
                         )
                     ),
@@ -817,36 +812,23 @@ class TableControlDialog(ui.Dialog, SettingsMixin):
         return self.settings.get("table_step_sizes") or self.default_steps
 
     def reset_position(self):
-        self.update_position(float('nan'), float('nan'), float('nan'))
+        self.update_position(Position())
 
-    def update_position(self, x, y ,z):
-        self.current_position = x, y, z
-        self.pos_x_label.value = x
-        self.pos_y_label.value = y
-        self.pos_z_label.value = z
+    def update_position(self, position):
+        self.current_position = position
+        self._position_widget.update_position(position)
         self.update_limits()
         self.update_control_buttons()
 
     def reset_caldone(self):
-        self.cal_x_label.value = None
-        self.cal_y_label.value = None
-        self.cal_z_label.value = None
-        self.rm_x_label.value = None
-        self.rm_y_label.value = None
-        self.rm_z_label.value = None
+        self._calibration_widget.reset_calibration()
 
-    def update_caldone(self, x, y ,z):
-        self.current_caldone = x, y, z
-        def getcal(value):
-            return value & 0x1
-        def getrm(value):
-            return (value >> 1) & 0x1
-        self.cal_x_label.value = getcal(x)
-        self.cal_y_label.value = getcal(y)
-        self.cal_z_label.value = getcal(z)
-        self.rm_x_label.value = getrm(x)
-        self.rm_y_label.value = getrm(y)
-        self.rm_z_label.value = getrm(z)
+    def update_caldone(self, position):
+        self.current_caldone = position
+        self.positions_widget.enabled = caldone_valid(position)
+        self.contacts_widget.enabled = caldone_valid(position)
+        self.control_layout.enabled = caldone_valid(position)
+        self._calibration_widget.update_calibration(position)
 
     def update_limits(self):
         x, y, z = self.current_position
@@ -970,11 +952,11 @@ class TableControlDialog(ui.Dialog, SettingsMixin):
         x, y, z = self.current_position
         callback(x, y, z)
 
-    def on_absolute_move(self, x, y, z):
-        z = safe_z_position(z)
+    def on_absolute_move(self, position):
+        z = safe_z_position(position.z)
         self.lock()
         self.stop_button.enabled = True
-        self.process.safe_absolute_move(x, y, z)
+        self.process.safe_absolute_move(position.x, position.y, position.z)
 
     def on_calibrate(self):
         self.lock()
@@ -1004,6 +986,10 @@ class TableControlDialog(ui.Dialog, SettingsMixin):
         self.positions_widget.load_settings()
         self.z_limit = settings.table_z_limit
         self.z_limit_label.value = self.z_limit
+        x, y, z = settings.table_probecard_maximum_limits
+        self.x_hard_limit_label.value = x
+        self.y_hard_limit_label.value = y
+        self.z_hard_limit_label.value = z
 
     def store_settings(self):
         self.settings['tablecontrol_dialog_size'] = self.width, self.height
@@ -1050,27 +1036,6 @@ class TableControlDialog(ui.Dialog, SettingsMixin):
             self.process.absolute_move_finished = None
             self.process.stopped = None
             self.process = None
-
-class CalibrationLabel(ui.Label):
-
-    def __init__(self, prefix, value=None):
-        super().__init__()
-        self.prefix = prefix
-        self.value = value
-
-    @property
-    def value(self):
-        return self.__value
-
-    @value.setter
-    def value(self, value):
-        self.__value = value
-        if value is None:
-            self.text = format(float('nan'))
-            self.qt.setStyleSheet("")
-        else:
-            self.text = f"{self.prefix} {self.__value}"
-            self.qt.setStyleSheet("QLabel:enabled{color:green}" if value else "QLabel:enabled{color:red}")
 
 class SwitchLabel(ui.Label):