+++ docs/changelog.md
@@ -7,6 +7,16 @@ nav_order: 10
 # Changelog
 {: .no_toc }
 
+## 0.30.0
+
+### Added
+- Restricted limits for joystick use.
+- Show points in plot line series.
+- Disable table controls if not calibrated.
+
+### Changed
+- Move position assignment into edit dialog.
+
 ## 0.29.2
 
 ### Changed